Wires, switches and more

+++ src/main/java/org/patryk3211/powergrid/electricity/creative/CreativeResistorBlockEntity.java
@@ -70,7 +70,7 @@ public class CreativeResistorBlockEntity extends ElectricBlockEntity implements 
     @Override
     protected void write(NbtCompound tag, boolean clientPacket) {
         super.write(tag, clientPacket);
-        tag.putFloat("Resistance", wire.getResistance());
+        tag.putFloat("Resistance", (float) wire.getResistance());
     }
 
     @Override
